Improve documentation and test structure
