+++ app/src/androidTest/java/com/google/samples/apps/topeka/activity/SignInActivityTest.java
@@ -147,8 +147,7 @@ public class SignInActivityTest {
     }
 
     private void editTextIsEmpty(int id) {
-        onView(withId(id))
-                .check(matches(withText(isEmptyOrNullString())));
+        onView(withId(id)).check(matches(withText(isEmptyOrNullString())));
     }
 
     @Test
