+++ app/src/androidTest/java/com/google/samples/apps/topeka/model/PlayerAndroidTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.topeka.helper;
+package com.google.samples.apps.topeka.model;
 
 import android.os.Parcel;
 import android.support.test.runner.AndroidJUnit4;
@@ -28,15 +28,23 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 @SmallTest
 @RunWith(AndroidJUnit4.class)
-public class ParcelableHelperTest {
+public class PlayerAndroidTest {
+
+    private static final Avatar AVATAR = Avatar.TWELVE;
+    private static final String LAST_INITIAL = "a";
+    private static final String FIRST_NAME = "first";
+
+    private static Player getPlayerUnderTest() {
+        return new Player(FIRST_NAME, LAST_INITIAL, AVATAR);
+    }
 
     @Test
-    public void writeReadBoolean() throws Exception {
-        Parcel testParcel = Parcel.obtain();
-        final boolean testValue = true;
-        ParcelableHelper.writeBoolean(testParcel, testValue);
-        testParcel.setDataPosition(0);
-        final boolean resultValue = ParcelableHelper.readBoolean(testParcel);
-        assertThat(testValue, is(resultValue));
+    public void writeToParcel() throws Exception {
+        Player initial = getPlayerUnderTest();
+        Parcel dest = Parcel.obtain();
+        initial.writeToParcel(dest, 0);
+        dest.setDataPosition(0);
+        Player unparcelled = new Player(dest);
+        assertThat(initial, is(unparcelled));
     }
 }
